alts: move ParseFramedMsg out of common
It's only called in one place, and is effectively a method on conn.

Part of #8510.
